[refactor] 경로 리팩토링
