+++ src/components/templates/detail/ContentSection/index.tsx
