+++ src/components/templates/home/PageInfoSection/index.tsx
