+++ src/pages/admin/dashboard/index.tsx
@@ -1,27 +1,26 @@
-import CommentLogTable from '@src/components/pages/admin/Dashboard/CommentLogTable';
-import FileLogTable from '@src/components/pages/admin/Dashboard/FileLogTable';
-import LoginLogTable from '@src/components/pages/admin/Dashboard/LoginLogTable';
-import PostLogTable from '@src/components/pages/admin/Dashboard/PostLogTable';
+import AdminCenterWrapper from '@src/components/common/Layout/AdminCenterWrapper';
+import CommentLogTable from '@src/components/templates/admin/Dashboard/CommentLogTable';
+import FileLogTable from '@src/components/templates/admin/Dashboard/FileLogTable';
+import LoginLogTable from '@src/components/templates/admin/Dashboard/LoginLogTable';
+import PostLogTable from '@src/components/templates/admin/Dashboard/PostLogTable';
 import { styled } from 'styled-components';
 
 function Dashboard() {
   return (
-    <Wrapper>
+    <CenterWrapper>
       <LoginLogTable />
       <FileLogTable />
       <PostLogTable />
       <CommentLogTable />
-    </Wrapper>
+    </CenterWrapper>
   );
 }
 
-const Wrapper = styled.div`
+const CenterWrapper = styled(AdminCenterWrapper)`
   display: flex;
   align-items: center;
   justify-content: center;
   flex-direction: column;
-  width: 800px;
-  margin: 100px auto;
   gap: 60px;
 `;
