+++ src/pages/admin/users/log/[page]/index.tsx
@@ -1,6 +1,6 @@
 import AdminCenterWrapper from '@src/components/common/Layout/AdminCenterWrapper';
 import BoardFooter from '@src/components/molecules/Board/BoardFooter';
-import UserLogTable from '@src/components/pages/admin/user/log/UserLogTable';
+import UserLogTable from '@src/components/templates/admin/user/log/UserLogTable';
 import { GetStaticPaths, GetStaticProps } from 'next';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
