+++ src/pages/boards/[board_type]/[post_id]/index.tsx
@@ -3,8 +3,8 @@ import PageTitle from '@src/components/common/PageTitle';
 import CommonCenterWrapper from '@src/components/common/Layout/CommonCenterWrapper';
 import { PATH } from '@src/utils/urls';
 import SideMenu from '@src/components/common/SideMenu';
-import CommentSection from '@src/components/pages/detail/CommentSection';
-import ContentSection from '@src/components/pages/detail/ContentSection';
+import CommentSection from '@src/components/templates/detail/CommentSection';
+import ContentSection from '@src/components/templates/detail/ContentSection';
 /**
  *  글 조회 페이지
  */
