+++ src/pages/boards/[board_type]/index.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import PageTitle from '@src/components/common/PageTitle';
 import SideBar from '@src/components/common/SideMenu';
-import Board from '@src/components/pages/boards/Board';
+import Board from '@src/components/templates/boards/Board';
 import PageWrapper from '@src/components/common/Layout/CommonCenterWrapper';
 import { PATH } from '@src/utils/urls';
 import { PAGE_TITLE } from '@src/utils/constants';
