+++ src/pages/mypage/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import PageTitle from '@src/components/common/PageTitle';
 import CommonCenterWrapper from '@src/components/common/Layout/CommonCenterWrapper';
-import ProfileForm from '@src/components/pages/mypage/ProfileForm';
+import ProfileForm from '@src/components/templates/mypage/ProfileForm';
 
 function MyPage() {
   return (
